+++ src/controllers/parkingspot.js
@@ -9,7 +9,7 @@ module.exports = {
   //PARKING SPOTS
   
   createParkingSpot: (cameraId, spotId, callback) => {
-    if(!cameraId || !spotId)
+    if (!cameraId || !spotId)
       return callback(new Error("Wrong parameters"), null)
     attemptConnection(`INSERT INTO parkingspot VALUES(${mysql.escape(cameraId)},${mysql.escape(spotId)},FALSE)`, (err,res) => {
       if (err) return callback(err,null)
@@ -21,7 +21,7 @@ module.exports = {
     })
   },
   getParkingSpotInfos: (cameraId, spotId, callback) => {
-    if(!cameraId || !spotId)
+    if (!cameraId || !spotId)
       return callback(new Error("Wrong parameters"), null)
     attemptConnection(`SELECT IsTaken FROM parkingspot WHERE CameraId = `+mysql.escape(cameraId)+` AND SpotNumber = `+mysql.escape(spotId), (err,res) => {
       if (err) return callback(err,null)
@@ -37,12 +37,12 @@ module.exports = {
     })
   },
   listCameraSpots: (cameraId,callback) => {
-    if(!cameraId)
+    if (!cameraId)
       return callback(new Error("Wrong parameters"), null)
     attemptConnection(`SELECT SpotNumber,IsTaken FROM parkingspot WHERE CameraId = `+mysql.escape(cameraId), (err, res) => {
-      if(err) return callback(err,null)
+      if (err) return callback(err,null)
       else {
-        if(!res[0]) {
+        if (!res[0]) {
           return callback(new Error("No parking spot exists for this camera"), null)
         } else {
           var list = []
@@ -61,7 +61,7 @@ module.exports = {
   },
   listParkingSpots: (callback) => {
     attemptConnection(`SELECT CameraId,SpotNumber,IsTaken FROM parkingspot`, (err, res) => {
-      if(err) return callback(err,null)
+      if (err) return callback(err,null)
       else {
         if(!res[0]) {
           return callback(new Error("No parking spot exists"), null)
@@ -81,8 +81,43 @@ module.exports = {
       }
     })
   },
+  listCamerasWithAvailableSpotsByAddress: (latitude,longitude,callback) => {
+    const MAXIMUM_DISTANCE_KM = 1
+    const MAXIMUM_NB_RESULT = 10
+    const EARTH_RADIUS_KM = 6371
+    attemptConnection(`SELECT C.CameraId,C.latitude,C.longitude, (COUNT(P.IsTaken) - SUM(P.IsTaken)) as nbAvailable,
+    ( ${EARTH_RADIUS_KM} * acos( cos( radians(${mysql.escape(latitude)}) ) * cos( radians( C.latitude ) ) * 
+    cos( radians( C.longitude ) - radians(${mysql.escape(longitude)}) ) + sin( radians(${mysql.escape(latitude)}) ) * 
+    sin( radians( C.latitude ) ) ) ) AS distance
+    FROM camera as C NATURAL JOIN parkingspot as P
+    GROUP BY C.CameraId
+    HAVING (COUNT(P.IsTaken) - SUM(P.IsTaken)) > 0 AND distance < ${MAXIMUM_DISTANCE_KM}
+    ORDER BY distance ASC
+    LIMIT 0,${MAXIMUM_NB_RESULT}`, (err,res) => {
+      if (err) return callback(err,null)
+      else {
+        if(!res[0]) {
+          return callback(new Error("No parking spot nearby"), null)
+        } else {
+          var list = []
+          res.forEach(row => {
+            list.push({
+              cameraId: row.CameraId,
+              latitude: row.latitude,
+              longitude: row.longitude,
+              nbAvailable: row.nbAvailable,
+              distance: row.distance
+            })
+          })
+          res.list = list
+          res.result = 'OK'
+          callback(err,res)
+        }
+      }
+    })
+  },
   updateParkingSpot: (cameraId, spotId, state, callback) => {
-    if(!cameraId || !spotId || !state)
+    if (!cameraId || !spotId || !state)
       return callback(new Error("Wrong parameters"), null)
     attemptConnection(`UPDATE parkingspot SET IsTaken = ${mysql.escape(state)} WHERE CameraId = `+mysql.escape(cameraId)+` AND SpotNumber = `+mysql.escape(spotId), (err,res) => {
       if (err) return callback(err,null)
@@ -94,7 +129,7 @@ module.exports = {
     })
   },
   deleteParkingSpot: (cameraId, spotId, callback) => {
-    if(!cameraId || !spotId)
+    if (!cameraId || !spotId)
       return callback(new Error("Wrong parameters"), null)
     attemptConnection(`DELETE FROM parkingspot WHERE CameraId = `+mysql.escape(cameraId)+` AND SpotNumber = `+mysql.escape(spotId), (err,res) => {
       if (err) return callback(err,null)